FastAPIBackend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
openai==1.10.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
twilio==8.12.0
python-dateutil==2.8.2
pytz==2024.1
fastapi
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pydantic
pydantic-settings
python-dateutil
python-dotenv
pytz
twilio
uvicorn[standard]
websockets
